fix: alterações finais, movendo a pasta para dentro do components e corrigindo a exportação no index de forma estruturada
