--- src/card/index.tsx
@@ -1,5 +1,0 @@
-export { Card } from './Card';
-export { CardBody } from './CardBody';
-export { CardHeader } from './CardHeader';
-export { CardImage } from './CardImage';
-export { CardProgress } from './CardProgress';
